# Optimize log injection with NLog