--- a/test/benchmarks/Benchmarks.Trace/NLogBenchmark.cs
+++ b/test/benchmarks/Benchmarks.Trace/NLogBenchmark.cs
@@ -14,13 +14,11 @@ namespace Benchmarks.Trace
     {
         private static readonly Tracer BaselineTracer;
         private static readonly Tracer LogInjectionTracer;
-        private static readonly NLog.Logger BaselineLogger;
-        private static readonly NLog.Logger EnrichedLogger;
+        private static readonly NLog.Logger Logger;
 
         static NLogBenchmark()
         {
-            var provider = new CustomNLogLogProvider();
-            provider.RegisterLayoutRenderers();
+            LogProvider.SetCurrentLogProvider(new CustomNLogLogProvider());
 
             var logInjectionSettings = new TracerSettings
             {
@@ -43,31 +41,23 @@ namespace Benchmarks.Trace
 
             BaselineTracer = new Tracer(baselineSettings, new DummyAgentWriter(), null, null, null);
 
+            var config = new LoggingConfiguration();
+
 #if DEBUG
             var writer = System.Console.Out;
 #else
             var writer = TextWriter.Null;
 #endif
 
-            var baselineConfig = new LoggingConfiguration();
-
-            baselineConfig.AddRuleForAllLevels(new TextWriterTarget(writer)
+            var target = new TextWriterTarget(writer)
             {
-                Layout = "${longdate}|${uppercase:${level}}|${logger}|{dd.env=,dd.service=,dd.version=,dd.trace_id=,dd.span_id=}|${message}"
-            });
-            
-            baselineConfig.LogFactory.ReconfigExistingLoggers();
-            
-            BaselineLogger = new LogFactory(baselineConfig).GetCurrentClassLogger();
+                Layout = "${longdate}|${uppercase:${level}}|${logger}|{dd.env=${mdlc:item=dd.env},dd.service=${mdlc:item=dd.service},dd.version=${mdlc:item=dd.version},dd.trace_id=${mdlc:item=dd.trace_id},dd.span_id=${mdlc:item=dd.span_id}}|${message}"
+            };
 
-            var enrichedConfig = new LoggingConfiguration(new LogFactory());
+            config.AddRuleForAllLevels(target);
 
-            enrichedConfig.AddRuleForAllLevels(new TextWriterTarget(writer)
-            {
-                Layout = $"${{longdate}}|${{uppercase:${{level}}}}|${{logger}}|{{dd.env=${{{CorrelationIdentifier.EnvKey}}},dd.service=${{{CorrelationIdentifier.ServiceKey}}},dd.version=${{{CorrelationIdentifier.VersionKey}}},dd.trace_id=${{{CorrelationIdentifier.TraceIdKey}}},dd.span_id=${{{CorrelationIdentifier.SpanIdKey}}}}}|${{message}}"
-            });
-
-            EnrichedLogger = new LogFactory(enrichedConfig).GetCurrentClassLogger();
+            LogManager.Configuration = config;
+            Logger = LogManager.GetCurrentClassLogger();
         }
 
         [Benchmark(Baseline = true)]
@@ -77,7 +67,7 @@ namespace Benchmarks.Trace
             {
                 using (BaselineTracer.StartActive("Child"))
                 {
-                    BaselineLogger.Info("Hello");
+                    Logger.Info("Hello");
                 }
             }
         }
@@ -89,7 +79,7 @@ namespace Benchmarks.Trace
             {
                 using (LogInjectionTracer.StartActive("Child"))
                 {
-                    EnrichedLogger.Info("Hello");
+                    Logger.Info("Hello");
                 }
             }
         }