Instead of registering the values of the properties for each span created, register properties for each trace and have those fetch the active span. 

|      Method |      Mean |     Error |    StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |----------:|----------:|----------:|-------:|------:|------:|----------:|
|         Old | 13.736 us | 0.0642 us | 0.0536 us | 0.1831 |     - |     - |  16.74 KB |
|         New |  5.899 us | 0.0828 us | 0.0774 us | 0.0687 |     - |     - |   6.34 KB |

Since Serilog and NLog now use a very similar logic, I think there's some refactoring to do. But I'm going to optimize log4net first so I can refactor them all at once.